feat: split middleware tests into their own files
